+++ R/mask_pixels.R
@@ -1,13 +1,14 @@
 #' Creates a mask over the background or foreground (Cria uma mascara sobre o
 #' background ou foreground)
 #'
-#' @description Esta funcao permite criar mascara sobre os pixels
-#'   correspondentes ao background ou foreground
+#' @description This function superimposes a mask over the pixels corresponding to
+#' the background or foreground.(Esta funcao permite criar mascara sobre os pixels
+#'   correspondentes ao background ou foreground)
 #' @usage mask_pixels(im,TargetPixels,TargetPixels2=NULL,col.TargetPixels="rand",
 #' Contour=FALSE, r = 1, g = 2, b = 3, norma = 1,plot=FALSE)
 
-#' @param im    :Este objeto deve conter uma imagem no formato do EBImage.
-#' @param TargetPixels    : Este objeto deve ser  uma matriz
+#' @param im    :This object must contain an image in EBImage format (Este objeto deve conter uma imagem no formato do EBImage).
+#' @param TargetPixels    : (Este objeto deve ser  uma matriz
 #'   binaria ou uma lista com varias matrizes. Em cada matriz deve conter os valores 0 (pixels do background) ou 1 (pixels do
 #'   foreground)).
 #' @param TargetPixels2    : Este objeto deve ser  uma matriz