+++ R/shape_file.BorderExtract.R
@@ -42,11 +42,16 @@ for(i in unique(ShapeFile[,1])){
 
 
   }
-
+A2=A
 if(!is.null(im)){
-plot_image(im)
-
-  for(i in unique(ShapeFile[,1])){
+  im2=resize_image(im,w=400)
+plot_image(im2)
+info1=info_image(im)$Length
+info2=info_image(im2)$Length
+A[,3]=info2[1]*A[,3]/info1[1]
+A[,4]=info2[2]*A[,4]/info1[2]
+ShapeFile=A
+for(i in unique(ShapeFile[,1])){
     id=A[,1]==i
     AA=A[id,]
 
@@ -65,5 +70,5 @@ plot_image(im)
   }
 }
 
-return(A)
+return(A2)
 }